Starting to flesh out the Monster protocol and records.
Decided to follow the book and have methods/fns write to println.
In the unit test I use with-out-str to redirect the output back
so it can be parsed and tested.
A fragile test strategy, but I'm learning some interesting features
of Clojure this way and not sure the best way to handle it.